expose notification type in CoreHome
