+++ plugins/CoreHome/vue/src/Notification/index.ts
@@ -10,3 +10,4 @@ import './Notifications.store.adapter';
 export { default as Notification } from './Notification.vue';
 export { default as NotificationGroup } from './NotificationGroup.vue';
 export { default as NotificationsStore } from './Notifications.store';
+export { default as NotificationType } from './Notification';
